refactor(plugins): rename user-list command to sidekick-options-container
