+++ bigbluebutton-html5/imports/ui/components/plugins-engine/ui-commands/handler.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import PluginChatUiCommandsHandler from './chat/handler';
-import PluginUserListUiCommandsHandler from './user-list/handler';
+import PluginSidekickOptionsContainerUiCommandsHandler from './sidekick-options-container/handler';
 
 const PluginUiCommandsHandler = () => (
   <>
     <PluginChatUiCommandsHandler />
-    <PluginUserListUiCommandsHandler />
+    <PluginSidekickOptionsContainerUiCommandsHandler />
   </>
 );
